Remove unused `phpexperts/laravel-env-polyfill`
The require of `phpexperts/laravel-env-polyfill` is opinionated and unnecessary for this package.
I'm removing the dependency because it is unused and causes problems for me:

The `phpexperts/laravel-env-polyfill` package can cause Laravel 8 TestSuites to fail / operate on wrong config because [server consts defined in `phpunit.xml`](https://github.com/laravel/laravel/blob/a6ca5778391b150102637459ac3b2a42d78d495b/phpunit.xml#L21-L29) are not being applied. This is an issue of `phpexperts/laravel-env-polyfill` on it's own. However, I'm removing this dependency here as a quick fix for myself since `phpexperts/laravel-env-polyfill` is not being used by `phpexperts/simple-dto` as far as I can tell.